Add confidence levels .csv and script to "narrow"
We manually created a data frame for confidence levels in each applicable variable for each data set in the open data portal (in order to easily see overlap between the same or similar variables between data sets). We then wrote a simple R script to display the results in a narrow data frame; the script and frame are uploaded as well.